% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNULISAseq.R
\name{add_plateID}
\alias{add_plateID}
\title{Add plateID column to a data frame within a list}
\usage{
add_plateID(x, element, plateID)
}
\arguments{
\item{x}{A list containing data frames as elements.}

\item{element}{Character. The name of the element in `x` to which the plateID should be added.}

\item{plateID}{Character or factor. The plate ID(s) to assign to the data frame.}
}
\value{
The data frame with an added or updated `plateID` column (as a factor).
}
\description{
This helper function adds or updates a `plateID` column in a specified data frame element within a list, ensuring the column is a factor. Used internally for merging NULISAseq data objects.
}
\keyword{internal}
