% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alamarColorPalette.R
\name{alamarColorPalette}
\alias{alamarColorPalette}
\title{Make color palette based on Alamar colors}
\usage{
alamarColorPalette(
  n,
  nReps = 1,
  tint = "lightdark",
  interpolate = FALSE,
  interpolateIndex = 1:5
)
}
\arguments{
\item{n}{number of main colors to output (maximum 13)}

\item{nReps}{optional, number of sub-colors to output 
for each of the main colors. Useful for when there are technical replicates.
Function will output n*nReps total colors.}

\item{tint}{optional, used when subcolors are created.
"light" blends main colors with white.
"dark" blends main colors with black.
"lightdark" (default) blends main colors with white & black.}

\item{interpolate}{logical TRUE or FALSE (default). 
If TRUE, main colors are based on interpolation of 
the colors denoted by the interpolate_index 
argument.}

\item{interpolateIndex}{indices of colors to interpolate. Default is 1:5}
}
\value{
A vector or list of vectors of colors in hex format
}
\description{
Make color palette based on Alamar colors
}
\examples{
cols <- alamarColorPalette(n=13, nReps=3)
plot(1:length(unlist(cols)),
1:length(unlist(cols)), col=unlist(cols),
pch=19, cex=4)

cols <- alamarColorPalette(n=10, nReps=3,
interpolate=TRUE, interpolateIndex=1:5)
plot(1:length(unlist(cols)),
1:length(unlist(cols)), col=unlist(cols),
pch=19, cex=4)


}
