% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBoxplot.R
\docType{data}
\name{colors_ICs}
\alias{colors_ICs}
\title{Sample Boxplot}
\format{
An object of class \code{character} of length 2.
}
\usage{
colors_ICs
}
\arguments{
\item{data_matrix}{The Data matrix output from readNULISAseq.R
or normalized data from normalization functions.}
}
\value{
Draws a plot.
}
\description{
Makes sample boxplots.
}
\keyword{datasets}
