% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_summary_stats.R
\name{column_summary_stats}
\alias{column_summary_stats}
\title{Generate Summary Statistics for Columns of a Matrix or Data Frame}
\usage{
column_summary_stats(x, n_digits = c(2, 2, 2, 3, 1), rowNames = colnames(x))
}
\arguments{
\item{x}{A matrix, data frame, or vector.}

\item{n_digits}{An optional vector of length 5 that indicates the number
of digits after the decimal place that will be used for rounding and 
formatting 
the mean, sd, median, min, and max, respectively. Default is
\code{c(2, 2, 2, 3, 1)]}. If NULL, no rounding or formatting will be done.}

\item{rowNames}{Names for the rows of output. Default will use 
column names of \code{x} as rows.}
}
\value{
A data frame with columns mean, sd, median, min, max, and 
number of missing values. Each row corresponds to the columns of the input.
}
\description{
Calculates mean, standard deviation, median, min, max, 
and number of missing values
for the columns of a matrix or for a vector.
}
