% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNULISAseq.R
\name{convert_to_long}
\alias{convert_to_long}
\title{A function that converts a matrix or data frame to long format by pivoting
sample columns into rows.}
\usage{
convert_to_long(
  data_matrix,
  id_col = "Target",
  sample_name_col = "SampleName",
  data_col
)
}
\arguments{
\item{data_matrix}{A matrix or data frame with IDs/targets as row names and
samples as columns.}

\item{id_col}{Character string specifying the name for the ID column
created from the row names (Default is "Target").}

\item{sample_name_col}{Character string specifying the name for the new
column that will contain the original sample column names (Default is "SampleName").}

\item{data_col}{Character string specifying the name for the values column
in the output (e.g., "Expression" or "Value"). This is a required argument.}
}
\value{
A data frame in long format with columns: ID, Sample Name, and Value.
}
\description{
A function that converts a matrix or data frame to long format by pivoting
sample columns into rows.
}
\keyword{internal}
