% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_AQ.R
\name{CV_AQ}
\alias{CV_AQ}
\title{Calculate Coefficient of Variation for Each Target for a Set of AQ Runs}
\usage{
CV_AQ(runs, exclude_outside_DR = TRUE)
}
\arguments{
\item{runs}{A named list of a single run data output from \code{laodNULISAseq()} 
function or a named list of these outputs for multiple runs.}

\item{exclude_outside_DR}{Logical. Default is TRUE. Should values outside the 
dynamic range (i.e., values below LLOQ or above ULOQ) be excluded 
before calculating CV?}
}
\value{
A nested list containing CV information. List elements 
include:
\item{run_intraCV}{A list with the same names as \code{runs}. Each run 
  has a sublist that includes a \code{intraCV} data frame with targets 
  in rows and samples as columns, and \code{n_samples} vector which gives
  number of sample replicates corresponding to each column in \code{intraCV}.}
  
\item{average_intraCV}{Summarizes intraCV across runs by taking the average 
  of the run-specific CVs. 
  This sublist also includes a \code{intraCV} data frame with targets 
  in rows and samples as columns, and \code{n_samples} vector which gives
  average number of sample replicates per run corresponding to each column 
  in \code{intraCV}. If only a 
  single run is input, \code{combined_intraCV} will be identical 
  to \code{run_intraCV}.}

\item{interCV}{If multiple runs are input, the \code{interCV} list is output.
  This gives the inter-plate CV calculated on the pooled replicates. Note this 
  is not a component inter-plate CV but a total CV across all runs. 
  This list includes a \code{interCV} data frame with targets 
  in rows and samples as columns, and \code{n_samples} vector which gives
  number of sample replicates corresponding to each column in \code{interCV}.}
}
\description{
Calculates intra-plate coefficient of variation (CV) 
for each target for a single AQ run or intra- and inter-plate CV for
a set of AQ runs. By default the function will output CVs for sample types 
SC (AQSC) and IPC (CAL). In the future, we might add the ability to also output 
CVs for samples that are grouped using a provided sample name pattern. 
Function input is the output of loadNULISAseq or a list
of loadNULISAseq outputs. CVs are calculated directly on the AQ data in 
aM units. Values outside of the dynamic range are excluded from CV calculation.
}
