% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_AQ_Hist.R
\name{CV_AQ_Hist}
\alias{CV_AQ_Hist}
\title{Make a CV Histogram for AQ Data}
\usage{
CV_AQ_Hist(CV_data, plot_type = NULL, title = NULL)
}
\arguments{
\item{CV_data}{A vector of CV values output from \code{CV_AQ()}
function.}

\item{plot_type}{Controls the tile of plot. Can be \code{'intra'} or 
{'inter'} for intra-plate or inter-plate CV, respectively. 
Default title will be "AQ Intra-plate CV (\%)" or "AQ Inter-plate CV (\%)".}

\item{title}{Optional title for the plot. Overrides default title.}
}
\value{
Generates a histogram plot.
}
\description{
Generates a coefficient of variation histogram using output of 
\code{CV_AQ()} function.
}
