% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvBoxplot.R
\name{cv_boxplot}
\alias{cv_boxplot}
\title{Coefficient of Variation Boxplot}
\usage{
cv_boxplot(cv_table, names, mar, title)
}
\arguments{
\item{cv_table}{Output of intraCV or interCV function.}
}
\value{
Draws a plot.
}
\description{
Makes boxplot of intra or inter-plate %CV by target.
}
