% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldChange.R
\name{FC}
\alias{FC}
\title{Observed vs Expected Fold Change}
\usage{
FC(interPlateNormData)
}
\value{
Draws a plot.
}
\description{
Calculates fold change across targets for pairs of samples. 
Option to calculate error from expected fold change values 
if those values are provided.
}
