% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcBoxplot.R
\name{fcBoxplot}
\alias{fcBoxplot}
\title{Fold Change Boxplot}
\usage{
fcBoxplot(FCdata, ylabel = NULL, ylimits = NULL, label_yval)
}
\arguments{
\item{fc_table}{Output of foldChange function.}
}
\value{
Draws a plot.
}
\description{
Makes boxplot of fold change or fold change error.
}
