% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQCplot.R
\name{filter_run_data}
\alias{filter_run_data}
\title{Filter Run Data for Each Plate}
\usage{
filter_run_data(
  run,
  sample_subset = NULL,
  sample_exclude = NULL,
  target_subset = NULL,
  target_exclude = NULL
)
}
\arguments{
\item{run}{A single plate or item from the named list of run data output from \code{loadNULISAseq()} function 
or a list of these outputs for multiple runs.}

\item{sample_subset}{Optional vector of sample names to include}

\item{sample_exclude}{Optional vector of sample names to exclude}

\item{target_subset}{Optional vector of target names to include}

\item{target_exclude}{Optional vector of target names to exclude}
}
\value{
Filtered run object
}
\description{
Internal function to filter sample and target data from each plate in the \code{runs} list.
Filters applied to \code{samples}, \code{targets}, \code{Data}, \code{lod$aboveLOD}, and \code{normed$interNormData} based on sample and target names.
}
\keyword{internal}
