% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQCplot.R
\name{generate_QCplot_metrics}
\alias{generate_QCplot_metrics}
\title{Generate QC Plot Metrics}
\usage{
generate_QCplot_metrics(runs)
}
\arguments{
\item{runs}{A named list of run data output from \code{loadNULISAseq()} function 
or a list of these outputs for multiple runs.}
}
\value{
List containing QC metrics and axis limits
}
\description{
Internal function to calculate QC plot axis limits for sample QC and target QC plots and generate QC summary.
}
\keyword{internal}
