% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNULISAseq.R
\name{get_internal_plate_id}
\alias{get_internal_plate_id}
\title{Extract AUTO_PLATE ID from a NULISA XML File}
\usage{
get_internal_plate_id(file_path)
}
\arguments{
\item{file_path}{Character string. The path to the XML file.}
}
\value{
A character string containing the plate ID if found; otherwise, returns `NA`.
}
\description{
Scans the beginning of an XML file to retrieve the value of the `AUTO_PLATE` 
attribute. This is used for early validation of plate identities before 
performing a full data load.
}
\details{
To optimize performance, the function only reads the first 1000 lines 
of the file, assuming metadata is located in the header section.
}
\keyword{internal}
