% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXML.R
\name{getXMLVersion}
\alias{getXMLVersion}
\title{Get XML Version from XML File}
\usage{
getXMLVersion(infile)
}
\arguments{
\item{infile}{The path to the XML file.}
}
\value{
A character string containing the XML version or "<1.3.0" if not found.
}
\description{
This function reads an XML file and returns the XMLversion value.
If the XMLversion node does not exist, it returns "<1.3.0".
}
\examples{
\dontrun{
# Example usage:
version <- getXMLVersion("your_file.xml")
}

}
