% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCV.R
\name{interCV}
\alias{interCV}
\title{Calculate Inter-plate Coefficients of Variation}
\usage{
interCV(
  data_list,
  samples,
  aboveLOD = NULL,
  exclude_targets = NULL,
  useMean = TRUE,
  method = "count"
)
}
\arguments{
\item{data_list}{A list of data matrices that contain data for 
the sample replicates for calculating inter-plate CV.}

\item{samples}{A list of vectors of length equal to 
the number of columns in the corresponding data matrix which encodes
the sets of sample replicates that will be used to calculate inter-plate CVs. 
Can be a list of either numeric or character vectors, where matching entries 
represent replicates for a given sample. 
Any NAs in the sample vectors will be excluded (it may be desirable to 
exclude negative controls or inter-plate controls, for example). 
Unique entries of these vectors will be used to identify 
unique sample replicate sets in the results output.}

\item{aboveLOD}{A list of logical TRUE/FALSE matrices output by the 
NULISAseqR::lod() function. Indicates which values to include / exclude 
in the CV calculation. 
Default is NULL, which includes all values in CV calculation.}

\item{exclude_targets}{A list of vectors of row names or row indices 
representing the 
targets that should be excluded from intra-plate CV calculation. For example,
one might want to exclude internal controls. Default is NULL, which includes
all common targets in the data_list matrices.}

\item{useMean}{A logical TRUE / FALSE that indicates whether or not 
to calculate inter-CV using the intra-plate means of replicates (TRUE, default), 
or to use the pooled replicate data (FALSE).}

\item{method}{Use "count" (default) for unnormalized or 
normalized count data. Calculates CV using the formula 
\code{cv = sd / mean}. Use "log2" for log2 transformed data. Calculates CV using 
the formula \code{sigma = sd(x) log(2)} and 
\code{cv = 100 sqrt(exp(sigma^2)-1)}.}
}
\value{
matrix with inter-plate %CV values for each target for each sample.
}
\description{
Calculates CV% for the specified replicates across multiple plates. 
This function only calculates inter-plate CV for 
targets that are present across all plates in the data_list. Any
targets that are present on only a subset of plates will be excluded.
In addition, individual targets, such as internal controls, can be 
explicitly excluded using the exclude_targets argument.
}
