% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interPlateNorm.R
\name{interPlateNormVector}
\alias{interPlateNormVector}
\title{Convenience function to apply inter-plate normalization on a vector}
\usage{
interPlateNormVector(data_list, ...)
}
\arguments{
\item{data_list}{A list of data matrices, one for each plate, 
to be normalized together.}

\item{...}{Other arguments to \code{interPlateNorm}. See \code{?interPlateNorm}.}
}
\value{
A list.
\item{interNormData}{A list of matrices of normalized count data (not 
log-transformed, unless input data was log-transformed, which should use `dataScale='log'`).}
\item{plate}{A vector of integers 1-(number of plates) 
corresponding to the interNormData matrix columns
that indicates which plate samples are from.} 
\item{normFactors}{A list of vectors with the normalization factor for each 
target in the corresponding plate}
\item{scaleFactor}{The scale factor that was used in the final scaling step.}
}
\description{
Output is a list.
}
\details{
Input is a list of IC-normalized data matrices, one for each plate. 
These would typically be normData matrix from the intraPlateNorm function 
or the Data matrix of `readNULISAseq.R` (if no intra-plate normalization is 
done). Other required input depends on methods used.
}
