% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmNULISAseq_predict.R
\name{lmNULISAseq_predict}
\alias{lmNULISAseq_predict}
\title{Linear Regression Model for NULISAseq Data - 
targets as predictors}
\usage{
lmNULISAseq_predict(
  data,
  sampleInfo,
  sampleName_var,
  response_var,
  modelFormula,
  reduced_modelFormula = NULL,
  exclude_targets = NULL,
  exclude_samples = NULL,
  target_subset = NULL,
  sample_subset = NULL,
  return_model_fits = FALSE
)
}
\arguments{
\item{data}{A matrix of normalized NULISAseq data
with targets used as predictors in rows, samples in columns. 
Row names should be the target names, and column names are the sample names.
It is assumed that data has already been transformed
using \code{log2(x + 1)} for each NULISAseq normalized count value \code{x}.}

\item{sampleInfo}{A data frame with sample metadata including the response 
variable and covariates. Rows are samples, columns are sample metadata variables. 
Linear regression models will only be done on the samples in sampleInfo, or a subset of those samples as 
specified using arguments \code{exclude_samples} or \code{sample_subset}. 
\code{sampleInfo} should have a column for each 
variable included in the linear regression models. String variables will 
be automatically treated as factors, and numeric variables will be 
treated as numeric.}

\item{sampleName_var}{The name of the column of sampleInfo that matches
the column names of \code{data}. This variable will be used to merge the 
target expression data with the sample metadata.}

\item{response_var}{The name of the column of sampleInfo specifying the continuous numeric response 
variable.}

\item{modelFormula}{A string that represents the right hand side of the model formula.
The main effect of target expression 
will be automatically added as a predictor. Any interactions need to be specified 
in the model formula as "covariate * target". 
For example \code{modelFormula = "disease + age + sex + plate"} tests for variations of outcome explained by the 
target predictor, adjusted for disease group, age, sex, and plate. \code{modelFormula = 
"disease * target + age + sex + plate"} includes both main and interaction 
effects for disease and target expression. See \code{?lm()}.}

\item{reduced_modelFormula}{Optional reduced model formula 
that contains only a subset of the terms in modelFormula. 
The reduced model serves as null model for an F-test using \code{anova()}. 
This could be useful for testing the overall significance of factor 
variables with more than 2 levels.}

\item{exclude_targets}{A vector of target names for targets that will be 
excluded from the linear regression models as predictors. Internal control targets, 
for example, should probably always be excluded.}

\item{exclude_samples}{A vector of sample names for samples that will be excluded
from the linear regression models. External control wells (IPCs, NCs, SC,)
should usually be excluded.}

\item{target_subset}{Overrides \code{exclude_targets}. A vector of target names 
for targets that will be included in the linear regression models as predictors.}

\item{sample_subset}{Overrides \code{exclude_samples}. A vector of sample names 
for samples that will be included in the linear regression models.}

\item{return_model_fits}{Logical \code{TRUE} or \code{FALSE} (default).
Should a list of the model fits be returned? Might be useful for more 
detailed analyses and plotting. However, also requires using more memory.}
}
\value{
A list including the following:
\item{modelStats}{A data frame with rows corresponding to targets and columns 
corresponding to estimated model coefficients, unadjusted p-values, 
Bonferroni adjusted p-values, and Benjamini-Hochberg false discovery rate
adjusted p-values (see \code{?p.adjust()}).}
\item{modelFits}{A list of length equal to number of targets containing
the model fit output from \code{lm()}. Only returned when 
\code{return_model_fits=TRUE}.}
\item{Fstats}{A data frame with rows corresponding to targets and columns.}
}
\description{
Fits linear regression model to each target in the NULISAseq data set,
using univariate targets as predictors in the model. 
Outputs coefficients, t-statistics, unadjusted and adjusted p-values.
This approach tests whether each 
target's expression is associated with the outcome variable, while 
adjusting for any specified covariates.
}
