% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod.R
\name{lod}
\alias{lod}
\title{Calculate Limits of Detection}
\usage{
lod(plate, blanks)
}
\arguments{
\item{data_matrix}{The Data matrix output from readNULISAseq.R
or normalized data from normalization functions.}
}
\description{
Calculates limit of detection (LoD) for each target based on the
negative controls. Designates data as either above or below LoD. 
Option to specify minimum count threshold for detectability.
}
