% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCplateHeatmap.R
\name{matrixify}
\alias{matrixify}
\title{Convert sample data to plate matrix format}
\usage{
matrixify(runSamples, plate = FALSE)
}
\arguments{
\item{runSamples}{Sample data from a run}

\item{plate}{Logical indicating whether to fill empty wells with ""}
}
\value{
Matrix representing the plate layout
}
\description{
Convert sample data to plate matrix format
}
\keyword{internal}
