% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULISAseqR.R
\docType{package}
\name{NULISAseqR-package}
\alias{NULISAseqR}
\alias{NULISAseqR-package}
\title{NULISAseqR: Analysis of NULISAseq Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides comprehensive analysis tools for NULISAseq proteomic data including data import, normalization, quality control, statistical analysis, and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nulisaseqr.alamarbio.com}
  \item \url{https://github.com/Alamar-Biosciences/NULISAseqR}
  \item Report bugs at \url{https://github.com/Alamar-Biosciences/NULISAseqR/issues}
}

}
\author{
\strong{Maintainer}: Alamar Biosciences Bioinformatics Team \email{bioinfo@alamarbio.com} (Contributors: Dwight Kuo, Joanne C. Beer, Eliza Chai, Sumedh Sankhe, Kasun Buddika) [copyright holder]

}
\keyword{internal}
