% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULISAseqR.R
\docType{package}
\name{NULISAseqR}
\alias{NULISAseqR}
\title{NULISAseqR Package}
\description{
NULISAseqR package can load data from Galaxy XML output files.
Includes functions to normalize, calculate CV, and visualize data, among others.
}
\author{
Joanne C. Beer \email{jbeer@alamarbio.com}
}
