% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{PCAplot}
\alias{PCAplot}
\title{Plot Principle Components}
\usage{
PCAplot(
  countMatrix,
  PCs = c(1, 2),
  colors = "black",
  labels = FALSE,
  label.colors = "black"
)
}
\arguments{
\item{data_matrix}{The Data matrix output from readNULISAseq.R
or normalized data from normalization functions.}
}
\value{
Draws a plot.
}
\description{
Makes scatterplot of samples using principal component scores.
}
