% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateHeatmap.R
\name{plateHeatmap}
\alias{plateHeatmap}
\title{Draw Plate Heatmap}
\usage{
plateHeatmap(data_matrix)
}
\arguments{
\item{data_matrix}{The Data matrix output from readNULISAseq.R
or normalized data from normalization functions.}
}
\value{
Draws a plot.
}
\description{
Makes heatmap for a specified target or total counts.
}
