% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{plot_plateLayout}
\alias{plot_plateLayout}
\title{Plot Plate Layout with Well Type Coloring}
\usage{
plot_plateLayout(
  runs,
  output_dir = NULL,
  plate_names = NULL,
  colors = NULL,
  col_width = "1.75cm",
  font_size = 8
)
}
\arguments{
\item{runs}{runs A named list of run data output from \code{loadNULISAseq()} function 
or a list of these outputs for multiple runs.}

\item{plate_names}{A vector of plate names, default is NULL and will be inferred from `runs`.}

\item{colors}{A vector of colors for different well types (IPC, NC, etc.), default is NULL to use a predefined color palette.}

\item{col_width}{The column width for the HTML table.}

\item{font_size}{The font size for the HTML table.}
}
\value{
A list of HTML elements with tables and plots for each plate layout.
}
\description{
This function generates plots and HTML tables for plate layouts, highlighting various well types (IPC, NC, SC, Bridge).
}
