% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXML.R
\name{processXML}
\alias{processXML}
\title{Write Processed NULISAseq XML}
\usage{
processXML(in_xml_file, IPC, NC, IC, barcodeB = NULL, out_XML = NULL)
}
\arguments{
\item{IPC}{Name to search for Interprocess control (IPC) samples}

\item{NC}{Name to search for Negative control (NC) samples}

\item{IC}{Name to search for Internal Control (IC) targets}

\item{barcodeB}{optional BarcodeB file}

\item{out_XML}{optional output XML to write to}

\item{xml_file:[file]}{Character string. Path and name of the file.}
}
\description{
Writes NULISAseq XML file with unnormalized / normalized data and QC flags
}
\examples{
# writeXML('filename.xml')

}
