% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNULISAseq.R
\name{protected_columns}
\alias{protected_columns}
\title{function to get the protected columns in the covariate data
This function returns a vector of column names that are protected and should not be cleaned
the column are created by ACC or upstream process and cannot be user edited
List of protected columns in NULISAseq sample data}
\usage{
protected_columns()
}
\value{
Character vector of protected column names.
}
\description{
This helper function returns a character vector of column names that should not be sanitized or altered during data cleaning and processing. These columns are essential identifiers or metadata fields in NULISAseq sample data.
}
\details{
- Used internally to prevent renaming or cleaning of key columns during sample data processing.
- Includes plate, sample, and experiment identifiers, as well as legacy and alternative naming conventions.
- Used Primarily in NAS for internal control of the samples dataframe
}
\keyword{internal}
