% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC2XML}
\alias{QC2XML}
\title{Write Processed XML from QC table}
\usage{
QC2XML(input, sample = F)
}
\arguments{
\item{input}{QC table}

\item{sample}{whether the table is sample or plate}
}
\value{
QC XML
}
\description{
Writes NULISAseq QC flags to XML
}
\examples{
# QC2XML(inputtable)

}
