% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCFlagSample}
\alias{QCFlagSample}
\title{Write Sample QC table}
\usage{
QCFlagSample(
  raw,
  aboveLOD,
  samples,
  targets,
  QCS = NULL,
  SN = NULL,
  well_order = NULL,
  ICs = NULL,
  IPCs = NULL,
  NCs = NULL,
  SCs = NULL,
  TAP = TRUE,
  xmlThresh = NULL,
  forceDefaults = FALSE
)
}
\arguments{
\item{raw}{unnormalized data}

\item{aboveLOD}{aboveLOD from the lod function}

\item{samples}{samples data frame}

\item{targets}{targets data frame}

\item{QCS}{QCS values}

\item{SN}{SN values}

\item{ICs}{name of IC samples}

\item{IPCs}{names of IPC samples}

\item{NCs}{names of NC samples}

\item{SCs}{names of SC samples}

\item{TAP}{whether this is a TAP run}

\item{xmlThresh}{Parsed XML QC thresholds (from readQCThresholdXMLNode) to override defaults, or NULL}

\item{forceDefaults}{If TRUE, ignore xmlThresh and use hardcoded defaults only}
}
\value{
QC table
}
\description{
Writes NULISAseq QC sample flags
}
\examples{
# QCFlagSample(inputtable)

}
