% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCPlateCriteria}
\alias{QCPlateCriteria}
\title{QCPlateCriteria}
\usage{
QCPlateCriteria(AQ = FALSE)
}
\arguments{
\item{AQ}{whether the experiment is an AQ experiment}
}
\value{
criteria
}
\description{
Plate QC Criteria
}
\examples{
# QCPlateCriteria()

}
