% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCTargetCriteria}
\alias{QCTargetCriteria}
\title{QCTargetCriteria}
\usage{
QCTargetCriteria(AQ = FALSE, advancedQC = FALSE)
}
\value{
criteria
}
\description{
Target QC Criteria
}
\examples{
# QCTargetCriteria()

}
