% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantHist.R
\name{quantHist}
\alias{quantHist}
\title{Make a Quantifiability Histogram}
\usage{
quantHist(quantifiability_data, group_name, title = NULL)
}
\arguments{
\item{quantifiability_data}{A list output from \code{quantifiability()}
function. The list has elements named \code{quant} which is a 
data frame with quantifiability values where targets are in rows and 
subgroups are in columns, and \code{n_samples} which is a named vector with 
sample sizes corresponding to the columns and column names of quant.}

\item{group_name}{The name of the subgroup for which to plot the data.
This name should match one of the columns of \code{quantifiability_data$quant}
and one of the entries of \code{quantifiability_data$n_samples}.}

\item{title}{Optional title. By default the title is generated using the 
group name and sample size.}
}
\value{
Generates a histogram plot.
}
\description{
Generates a quantifiability histogram using output of 
\code{quantifiability()} function.
}
