% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNULISAseq.R
\name{readCovariateFile}
\alias{readCovariateFile}
\title{Read Covariate file and add to NULISAseq object}
\usage{
readCovariateFile(txt_file, NULISAseqRuns)
}
\arguments{
\item{txt_file}{Character string. File containing covariates (columns with name not already existing}

\item{NULISAseqRuns}{List of NULISAseq run objects}
}
\value{
List of lists, data frames, and matrices.
}
\description{
Reads NULISAseq covariate file and adds new covariates to list object. Merge based on plateID, row, col and sampleName. 
Fails if plateID is not set AND more than one entry in list
}
