% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXML.R
\name{removeBarcodeAndSamples}
\alias{removeBarcodeAndSamples}
\title{Removes BarcodeB nodes with a specified attribute and their associated Sample nodes.}
\usage{
removeBarcodeAndSamples(doc, attribute_name, attribute_value)
}
\arguments{
\item{doc}{An XML document.}

\item{attribute_name}{The name of the attribute to match in BarcodeB nodes.}

\item{attribute_value}{The value of the attribute to match in BarcodeB nodes. Can be multiple values.}
}
\value{
The modified XML document with BarcodeB and associated Sample nodes removed.
}
\description{
This function takes an XML document, along with the name and value of an attribute,
and removes the BarcodeB nodes that match the specified attribute. Additionally,
it identifies the associated Sample nodes with the same name attribute as the removed
BarcodeB nodes and removes them as well. These removals only occur if the sample/barcode
node in question is not a control (e.g. IPC, NC, SC, Bridge, etc.)
}
\examples{
doc <- read_xml("your_file.xml")
new_doc <- removeBarcodeAndSamples(doc, "name", c("774951", "23112"))

}
