% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNULISAseq.R
\name{rename_cols}
\alias{rename_cols}
\title{A function to rename the column names of a matrix given a dataframe of old and new
names}
\usage{
rename_cols(mat, names_df)
}
\arguments{
\item{mat}{A named (rownames/column names) matrix}

\item{names_df}{A two column dataframe with sampleName and sampleID as the old and new column name
respectively}
}
\value{
A matrix with the updated columns names
}
\description{
A function to rename the column names of a matrix given a dataframe of old and new
names
}
\keyword{internal}
