% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNULISAseq.R
\name{renameSC}
\alias{renameSC}
\title{Rename Sample Controls (SC) and Absolute Quantification Sample Controls (AQSC)}
\usage{
renameSC(column, label = "SC", label2 = "AQSC", n = 3)
}
\arguments{
\item{column}{A character vector or column of values to be renamed. It should contain
"SC" and "AQSC" values that need to be renamed.}

\item{label}{A string specifying the first label to be renamed (default is "SC").}

\item{label2}{A string specifying the second label to be renamed (default is "AQSC").}

\item{n}{An integer specifying the maximum value for the counter. The counter cycles
from 1 to `n` for each label occurrence (default is 3).}
}
\value{
A character vector with the renamed values. The labels "SC" and "AQSC" will
        have a counter appended to them (e.g., "SC_1", "AQSC_1", etc.), while other
        values will remain unchanged.
}
\description{
This function renames instances of "SC" and "AQSC" in the input column by appending
a counter to each occurrence. The counter cycles through values from 1 to `n`,
and resets back to 1 after reaching the specified limit.
}
\details{
The function works by iterating over the input vector and checking each value.
         If the value is equal to `label` or `label2`, a counter is appended to it.
         The counter resets to 1 after reaching the value `n`. This is useful for ensuring
         unique naming of standard controls and assay quality standard controls in a dataset.
}
\examples{
# Example usage:
renameSC(c("SC", "AQSC", "SC", "Other", "AQSC"))
renameSC(c("SC", "AQSC", "SC", "Other", "AQSC"), n = 2)

}
\keyword{internal}
