% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghettiPlot.R
\name{spaghetti_plot}
\alias{spaghetti_plot}
\title{Spaghetti plot for visualizing longitudinal NULISAseq data}
\usage{
spaghetti_plot(
  target,
  data,
  sampleInfo,
  time_var_name,
  id_var_name,
  group_var_name = NULL,
  axis_label_font_size = 0.55,
  legend = TRUE
)
}
\arguments{
\item{axis_label_font_size}{Font size of axis title.}

\item{coefs}{Vector of coefficients.}

\item{p_vals}{Vector of p-values.}

\item{target_labels}{Vector of target names.}

\item{title}{Plot title.}

\item{xlabel}{X axis label.}

\item{ylabel}{Y axis label.}

\item{xlimits}{X axis limits}

\item{ylimits}{Y axis limits}

\item{max.overlaps}{Passed to ggrepel. Integer which determines how many targets 
will have labels. "Inf" (default) labels all targets.}

\item{plot_title_font_size}{Font size of plot title.}

\item{tick_label_font_size}{Font size of axis tick marks.}

\item{target_label_size}{Font size of target labels.}

\item{target_label_segment_color}{Color of line segments for target labels.}

\item{plot_aspect_ratio}{Aspect ratio for plot. Default is 1 (square plot).}
}
\value{
Outputs a ggplot object.
}
\description{
Draws a volcano plot for a set of coefficients and p-values. Uses ggplot
for plotting and ggrepel package for labels. By default, plot labeling
assumes FDR-corrected p-values.
}
