% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXML.R
\name{splitXML}
\alias{splitXML}
\title{Split XML file based on attribute and create a ZIP archive}
\usage{
splitXML(xmlFile, attr_name2)
}
\arguments{
\item{xmlFile:file}{Character string vector. Path(s) and name(s) of the file(s).}

\item{attr_name2:[str]}{The attribute name used to identify nodes for splitting.}
}
\description{
This endpoint reads an XML file, identifies specific elements based on an attribute name,
removes nodes with the specified attribute value, creates new XML files, and zips them together.
}
