% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetBoxplot.R
\name{targetBoxplot}
\alias{targetBoxplot}
\title{Target Boxplot}
\usage{
targetBoxplot(data_matrix)
}
\arguments{
\item{data_matrix}{The Data matrix output from readNULISAseq.R
or normalized data from normalization functions.}
}
\value{
Draws a plot.
}
\description{
Makes boxplot of target counts.
}
