% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetQCplot.R
\name{targetQCplot}
\alias{targetQCplot}
\title{Generate a Target QC Plot for AQ Panel}
\usage{
targetQCplot(runs)
}
\arguments{
\item{runs}{A list of run data output from \code{importNULISAseq()$runs}
or a list of the outputs from \code{loadNULISAseq}.}
}
\value{
Generates a target QC plot.
}
\description{
Generates a recovery plot for absolute quantification (AQ) data sample 
controls based on their expected vs observed concentration.
}
