% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateSummary.R
\name{typeSummary}
\alias{typeSummary}
\title{Well Type Summary}
\usage{
typeSummary(well_type, plate_data, total_plate_reads)
}
\description{
Helper function to summarize special well types:
IPCs
NCs
SCs
Bridge samples
}
