% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCplateHeatmap.R
\name{wellcolors}
\alias{wellcolors}
\title{Assign colors to wells based on their type}
\usage{
wellcolors(data, well_types, boxplot_colors)
}
\arguments{
\item{data}{Data matrix with well names as column names}

\item{well_types}{List of well type patterns to match}

\item{boxplot_colors}{Vector of colors to assign}
}
\value{
Vector of colors for each well
}
\description{
Assign colors to wells based on their type
}
\keyword{internal}
