% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{withinSampleTest}
\alias{withinSampleTest}
\title{Differential Expression Test for all Targets.}
\usage{
withinSampleTest(interPlateNormData)
}
\value{
Draws a plot.
}
\description{
Within or between-sample t-test.
}
