% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNULISAseq.R
\name{writeNULISAseq}
\alias{writeNULISAseq}
\title{Write NULISAseq normalized data CSV in long format}
\usage{
writeNULISAseq(
  xml_files,
  dataDir,
  target_info_file,
  sample_info_file = NULL,
  sample_info_file_variables = NULL,
  output_filename,
  Panel = "200-plex Inflammation v1",
  PanelLotNumber = "",
  plateIDs = NULL,
  ICs = "mCherry",
  IPC_string = "IPC",
  SC_string = "SC",
  Bridge_string = "Bridge",
  NC_string = "NC",
  include_IPC = TRUE,
  include_SC = TRUE,
  include_Bridge = TRUE,
  include_NC = FALSE,
  intraPlateNorm_method = "IC",
  intraPlateNorm_scaleFactor = 1,
  interPlateNorm_method = "IPC",
  IPC_method = "median",
  IN_samples = NULL,
  interPlateNorm_dataScale = "count",
  interPlateNorm_scaleFactor = 1,
  replaceNA = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{xml_files}{Vector of filenames (character strings). Files should 
be in order of the desired plateID variable, unless that is otherwise
defined.}

\item{dataDir}{Data directory where xml_files, target_info_file, and sample_info_file 
reside (character string).}

\item{target_info_file}{Path and filename for the target info CSV file 
(character string). Must include columns for TargetName (matches the targets in 
XML files), AlamarTargetID, UniProtID, and ProteinName. Only targets in the 
target_info_file will be output in the CSV file.}

\item{sample_info_file}{Optional. Path and filename for the sample annotation CSV file. 
Must include columns "plateID" and "sampleName" which correspond to the 
plateID (matching the plateIDs input to this function) and sampleName in the readNULISAseq 
samples data.frame.}

\item{sample_info_file_variables}{Subset of column names in 
sample_info_file that will be included in data file output. Other columns 
will be excluded. Otherwise, if NULL (default), all columns will be included.}

\item{output_filename}{Filename for output CSV file.}

\item{Panel}{Name of multi-plex panel. Default is '200-plex Inflammation v1'}

\item{PanelLotNumber}{The panel lot number.}

\item{plateIDs}{A vector of plate IDs. If NULL, default is to number plates from 01 to 
total number of plates based on the order of xml_files. Passed to readNULISAseq() function
and output as a column in the data file.}

\item{ICs}{vector of string(s). Internal control names. Default is "mCherry". 
First IC in vector will be used in intra-plate IC-normalization (usually mCherry). 
ICs will be omitted from CSV output by default. 
If NULL, no intra-plate normalization is done.}

\item{IPC_string}{IPC_string (or IPC_wells) is required for IPC and IN normalization.
Vector of character string(s) that represents IPCs in the column names (e.g. 'IPC').}

\item{SC_string}{Optional (or SC_wells) Vector of character string(s) that represents SCs in the column 
names (e.g. 'SC'). SC wells are included in the data file output.}

\item{Bridge_string}{Optional (or Bridge_wells) Vector of character string(s) that represents 
Bridge samples in the column names (e.g. 'Bridge'). Bridge wells 
can be used for inter-plate normalization (not currently implemented).}

\item{NC_string}{Required (or NC wells). Vector of character string(s) that represents NCs in the 
column names (e.g. 'NC').}

\item{include_IPC}{Logical. Should IPC samples be included in output? Default is TRUE.}

\item{include_SC}{Logical. Should SC samples be included in output? Default is TRUE.}

\item{include_Bridge}{Logical. Should Bridge samples be included in output? Default is TRUE.}

\item{include_NC}{Logical. Should NC samples be included in output? Default is FALSE.}

\item{intraPlateNorm_method}{intra-plate normalization method passed to 
intraPlateNorm() function. Default is 'IC' (internal control). Other option is
'TC' (total count), but this method is not currently implemented.}

\item{intraPlateNorm_scaleFactor}{Optional scaling factor to apply after 
intra-plate normalization. Passed to intraPlateNorm() function.}

\item{interPlateNorm_method}{Default is "IPC" for inter-plate control normalization. 
Use "IN" for intensity normalization. If neither, no interplate normalization is done.}

\item{IPC_method}{Passed to interPlateNorm function. 'median' is the default. 
Other options include 'mean' (arithmetic mean) and 'geom_mean' (geometric mean). 
Determines how the counts are summarized across the IPC wells on a given plate.}

\item{IN_samples}{Optional argument. Passed to interPlateNorm function. 
A list of column names or indices specifying which subset of samples to use 
for intensity normalization step for each plate in data_list. Will over-ride 
the IPC_wells and NC_wells arguments for IN.}

\item{interPlateNorm_dataScale}{Passed to interPlateNorm function. 'count' is 
the default and interplate normalization is multiplicative. Use option 'log' 
for log-transformed data; normalization is additive on the log scale.}

\item{interPlateNorm_scaleFactor}{Passed to interPlateNorm function. Optional 
numeric value used to rescale all data after normalizing. Default is 1. 
This may be desirable to avoid normalized quantities between 0 and 1 
(which will be negative in the log scale). Only useful for count scale data.}

\item{replaceNA}{Logical. Passed to readNULISAseq() function.
If TRUE (default), will replace missing counts with 
zeros.}

\item{verbose}{Logical. Should function output step completion into.
Default is TRUE.}
}
\value{
writes CSV file
}
\description{
Reads in one or more NULISAseq XML files, where XML file is output from the Alamar
Biosciences Galaxy NULISAseq (Beta) tool. Output is a long-format CSV file
where each row corresponds to a particular sample-target combination.
}
\details{
This function performs intra-plate 
normalization (default is internal control (IC) mCherry), and when multiple 
XML files are input, also performs inter-plate normalization (default is inter-plate 
control (IPC)). Finally, data is log2-transformed (after adding small constant, 0.01). 

The function also takes as input a target info file containing target metadata, 
including Alamar target IDs and protein names, and an optional sample info 
file containing sample metadata that
is desired to appear in the output file.

Sample QC is currently based on the internal control (mCherry) within 30% of
median criterion. If a sample's mCherry count falls outside this threshold, 
it will be assigned a "WARN" status. Otherwise the sample will be assigned a 
"PASS" status.  

The function outputs a CSV file where each row corresponds to a particular
sample-target combination. Total number of rows is number of samples times
number of targets. By default the sample control (SC) and inter-plate control 
(IPC) wells are included in the output, and NC wells are omitted.


Columns of output data (in order) include:
\itemize{
 \item{"PanelLotNumber"}{}
 \item{"SampleID"}{}
 \item{"(optional sample metadata columns)"}{}
 \item{"Target"}{}
 \item{"AlamarTargetID"}{}
 \item{"UniProtID"}{}
 \item{"ProteinName"}{}
 \item{"SampleQC"}{}
 \item{"LOD"}{}
 \item{"log2NormalizedCount"}{}
}
}
