% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXML.R
\name{writeUpdatedXML}
\alias{writeUpdatedXML}
\title{Write Updated XML}
\usage{
writeUpdatedXML(input_XML, rawReadCutoff = 200, data = NULL)
}
\arguments{
\item{input_XML}{Path to the input XML file to be processed}

\item{rawReadCutoff}{The raw read value cutoff to apply to individual sample/target combinations}

\item{data}{output of loadNULISAseq (optional)}
}
\value{
XML document object containing the NULISAseq data
}
\description{
Reads and processes an XML file containing NULISAseq data. This function reads the input XML file
and returns the XML document with updated info (v1.3.0) from loadNULISAseq (e.g. AQ and QC info).
}
\examples{
# writeUpdatedXML("path/to/input.xml")

}
