% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alamarColorPalette.R
\name{alamarColorPalette}
\alias{alamarColorPalette}
\title{Make color palette based on Alamar colors}
\usage{
alamarColorPalette(
  n,
  nReps = 1,
  tint = "lightdark",
  interpolate = FALSE,
  interpolateIndex = NULL,
  palette = 1
)
}
\arguments{
\item{n}{number of main colors to output. Maximum is 13 when interpolate=FALSE.}

\item{nReps}{optional, number of sub-colors to output 
for each of the main colors. Useful for when there are technical replicates.
Function will output n*nReps total colors.}

\item{tint}{optional, used when subcolors are created.
"light" blends main colors with white.
"dark" blends main colors with black.
"lightdark" (default) blends main colors with white & black.}

\item{interpolate}{logical TRUE or FALSE (default). 
If TRUE, main colors are based on interpolation of 
the colors denoted by the interpolate_index 
argument.}

\item{interpolateIndex}{indices of colors to interpolate. For `palette=1`, 
default is 1:5. For `palette=2`, default is `c(1,2,4,3,5,6)`.}

\item{palette}{Which set of colors to use. Integer. `"palette=1"` (default) 
is based on the original website color palette. `palette=2` is the new 
palette used in the NULISA manuscript.}
}
\value{
A vector or list of vectors of colors in hex format
}
\description{
Make color palette based on Alamar colors
}
\examples{
cols <- alamarColorPalette(n=13, nReps=3)
plot(1:length(unlist(cols)),
1:length(unlist(cols)), col=unlist(cols),
pch=19, cex=4)

cols <- alamarColorPalette(n=10, nReps=3,
interpolate=TRUE, interpolateIndex=1:5)
plot(1:length(unlist(cols)),
1:length(unlist(cols)), col=unlist(cols),
pch=19, cex=4)


}
