% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectability.R
\name{detectability}
\alias{detectability}
\title{Calculate Detectability for a Set of Targets and Samples}
\usage{
detectability(aboveLOD_matrix, sample_subset = NULL, exclude_targets = NULL)
}
\arguments{
\item{aboveLOD_matrix}{aboveLOD matrix output from lod() function.
Rows are targets and columns are samples.}

\item{sample_subset}{A vector of column names or numeric column indices that
represent the sample subset that detectability will be calculated for. 
Default uses all columns. NCs and IPCs should probably be excluded.}

\item{exclude_targets}{A vector of row names or numeric row indices 
representing the 
targets that should be excluded from detectability calculation. For example,
one might want to exclude internal controls. Default is NULL, which includes
all targets in the aboveLOD_matrix.}

\item{detectability}{A named vector of detectability values as percentages. 
Names correspond to target row names.}

\item{detectable}{A named TRUE/FALSE vector that indicates whether 
detectability is above 50% for that target. 
Names correspond to target row names.}
}
\value{
A list.
}
\description{
Calculates detectability for each target based on the
aboveLOD matrix output of lod() function. Detectability is percent of samples
above LOD for that target.
}
