% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectability.R
\name{detectability_summary}
\alias{detectability_summary}
\title{Summarize Detectability Across Multiple Runs and Sample Groups}
\usage{
detectability_summary(runs)
}
\arguments{
\item{runs}{A named list of run data output from \code{laodNULISAseq()} function 
or a list of these outputs for multiple runs. To make output more interpretable,
it is recommended to name each run according to the plate ID for that run.}
}
\value{
A list with items "sample_group", if applicable, "all", and "run_detectability".
Items sample_group and all each have a data frame with a row for each target 
and detectability in columns. "all" has only one column summarizing detectability for all 
samples. "sample_group" has a column summarizing detectability for each 
sample group. Each of there items also has a sampleNumber vector which gives the 
sample size for the groups. The run_detectability output is a list of the 
original individual run detectability objects from loadNULISAseq.
}
\description{
Summarizes detectability across multiple runs and if applicable 
broken down by sample groups based on the sample_group_covar input
to loadNULISAseq.
}
