% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXML.R
\name{insertCovariatesXML}
\alias{insertCovariatesXML}
\title{Inserts Covariates into XML}
\usage{
insertCovariatesXML(input_XML, covariates)
}
\arguments{
\item{input_XML}{Path to the input XML file.}

\item{covariates}{A data frame containing covariate information, including 'sampleName' 
and other covariate columns.}
}
\value{
The modified XML document.
}
\description{
This function reads an XML file, modifies specific attributes within the <Barcode> elements 
under <BarcodeB>, and inserts corresponding covariate values from a provided data frame.
}
\examples{
# Example usage:
# data <- loadNULISAseq('input.xml')
# data$samples$newAnnot <- c(1:length(1:nrow(data$samples)))
# insertCovariatesXML("input.xml", data$samples)

}
