% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCV.R
\name{interCV}
\alias{interCV}
\title{Calculate Inter-plate Coefficients of Variation}
\usage{
interCV(norm_data, techReps, techRepOrder_alpha, techRepOrder_kf)
}
\arguments{
\item{plates}{A list of count data matrices.}
}
\value{
matrix with %CV values for each target for each sample.
}
\description{
Calculates CV% for the specified replicates across multiple plates.
}
