% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interPlateNorm.R
\name{interPlateNorm}
\alias{interPlateNorm}
\title{NULISAseq Inter-Plate Normalization}
\usage{
interPlateNorm(
  data_list,
  IPC = TRUE,
  IN = FALSE,
  IPC_wells = NULL,
  NC_wells = NULL,
  IPC_method = "median",
  IN_samples = NULL,
  dataScale = "count",
  scaleFactor = 1
)
}
\arguments{
\item{data_list}{A list of data matrices, one for each plate, 
to be normalized together.}

\item{IPC}{Logical TRUE/FALSE indicating whether a IPC 
normalization should be done. Default is TRUE.}

\item{IN}{Logical TRUE/FALSE indicating whether intensity normalization
should be done. Recommended only when samples are randomized 
across the plates. If targets do not match across all plates, only the 
matching targets will be returned in the output. Default is FALSE.}

\item{IPC_wells}{List of vectors of either the column indices (numeric) 
or the column names (character string) of the inter-plate control wells 
for each plate. IPC wells are omitted from the 
intensity normalization median calculation. 
Required for IPC normalization and recommended for IN.}

\item{NC_wells}{Recommended for intensity normalization. 
List of vectors of either the row indices (numeric) or the 
row names (character string) 
of the negative control wells for each plate. NC wells are omitted from 
intensity normalization.}

\item{IPC_method}{'median' is the default. Other options include 
'mean' (arithmetic mean) and 'geom_mean' (geometric mean). 
Determines how the counts are summarized 
across the IPC wells on a given plate.}

\item{IN_samples}{Optional argument. A list of column names or
indices specifying which subset of samples to use for 
intensity normalization step for each plate in data_list. 
Will over-ride the IPC_wells and 
NC_wells arguments for IN.}

\item{dataScale}{'count' is the default and interplate 
normalization is multiplicative. Use option 'log' for log-transformed
data; normalization is additive on the log scale.}

\item{scaleFactor}{Optional numeric value used to rescale all data 
after normalizing. Default is 1. This may be desirable to avoid
normalized quantities between 0 and 1 (which will be negative
in the log scale). Only useful for count scale data.}

\item{interNormData}{A matrix of normalized count data (not log-transformed).}

\item{plate}{A vector of integers 1-(number of plates) 
corresponding to the interNormData matrix columns
that indicates which plate samples are from.}
}
\value{
A list.
}
\description{
Does inter-plate normalization using one of the following methods:
inter-plate control (IPC; divides each target 
by the median IPC count for that target), intensity normalization (IN; 
sets median of each target on each plate to equal the global median). Default
is to do IPC. 
  
Input is a list of data matrices, one for each plate. 
These would typically be normData matrix from the intraPlateNorm function 
or the Data matrix of readNULISAseq.R (if no intra-plate normalization is 
done). Other required input depends on methods used.
}
\details{
Use option dataScale = log when data is already on a log scale.

Output is a single matrix.
}
