% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraCV.R
\name{intraCV}
\alias{intraCV}
\title{Calculate Intra-plate Coefficients of Variation}
\usage{
intraCV(countMatrix, samples)
}
\arguments{
\item{data_matrix}{The Data matrix output from readNULISAseq.R
or normalized data from normalization functions.}
}
\value{
matrix with %CV values for each target for each 
multiple-replicate sample.
}
\description{
Calculates CV% for the specified replicates on a given plate.
}
