% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraCV.R
\name{intraCV}
\alias{intraCV}
\title{Calculate Intra-plate Coefficients of Variation}
\usage{
intraCV(data_matrix, samples, aboveLOD = NULL, exclude_targets = NULL)
}
\arguments{
\item{data_matrix}{The Data matrix output from readNULISAseq.R
or normalized data from normalization functions.}

\item{samples}{A vector of length equal to ncol(data_matrix) which encodes
the sets of sample replicates that will be used to calculate intra-plate CVs. 
Can be either a numeric or a character vector, where matching entries 
represent replicates for a given sample. 
Any NAs will be excluded (it may be desirable to 
exclude negative controls or inter-plate controls, for example). 
Entries of this vector will be used to identify unique sample replicate sets 
in the results output.}

\item{aboveLOD}{Logical TRUE/FALSE matrix output by the NULISAseqR::lod()
function. Indicates which values to include / exclude in the CV calculation. 
Default is NULL, which includes all values in CV calculation.}

\item{exclude_targets}{A vector of row names or row indices representing the 
targets that should be excluded from intra-plate CV calculation. For example,
one might want to exclude internal controls. Default is NULL, which includes
all targets in the data_matrix.}

\item{cv_matrix}{matrix with %CV values for each target for each 
multiple-replicate sample.}
}
\value{
A list.
}
\description{
Calculates CV% for multiple targets 
for the specified sample replicates on a given plate.
}
