% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerNULISAseq.R
\name{lmerNULISAseq}
\alias{lmerNULISAseq}
\title{Linear regression model for NULISAseq differential expression test}
\usage{
lmerNULISAseq(
  data,
  sampleInfo,
  sampleName_var,
  modelFormula_fixed,
  modelFormula_random,
  reduced_modelFormula_fixed = NULL,
  reduced_modelFormula_random = NULL,
  exclude_targets = NULL,
  exclude_samples = NULL,
  target_subset = NULL,
  sample_subset = NULL,
  return_model_fits = FALSE
)
}
\arguments{
\item{data}{A matrix of normalized NULISAseq data
with targets in rows, samples in columns. 
Row names should be the target names, and column names are the sample names.
It is assumed that data has already been transformed
using \code{log2(x + 0.01)} for each NULISAseq normalized count value \code{x}.}

\item{sampleInfo}{A data frame with sample metadata. Rows are samples, 
columns are sample metadata variables. Differential expression analysis will 
only be done on the samples in sampleInfo, or a subset of those samples as 
specified using arguments \code{exclude_samples} or \code{sample_subset}. 
\code{sampleInfo} should have a column for each 
variable included in the linear mixed effect models. String variables will 
be automatically treated as factors, and numeric variables will be 
treated as numeric.}

\item{sampleName_var}{The name of the column of sampleInfo that matches
the column names of \code{data}. This variable will be used to merge the 
target expression data with the sample metadata.}

\item{modelFormula_fixed}{A string that represents the fixed effects part of the model 
formula on the used for the linear mixed effects model. 
For example \code{modelFormula_fixed = 
"disease + age + sex + plate"} tests for differences in target expression 
by disease group, adjusted for age, sex, and plate. \code{modelFormula = 
"disease * age + sex + plate"} includes both main and interaction 
effects for disease and age. See \code{?lmer()}.}

\item{modelFormula_random}{A string that represents the random effects part of the model 
formula on the used for the linear mixed effects model. 
For example \code{modelFormula_random = "(1|participant_ID)"} 
creates a subject specific random intercept, where the variable 
\code{participant_ID} (a column in \code{sampleInfo} data frame) denotes 
repeated measures on the same subject. For subject-specific random intercept and 
slopes (not recommended when time is categorical), 
use \code{modelFormula_random = "(1 + time|participant_ID)"}. For random 
subject nested within plate (which may be useful when analyzing
a large number of plates together), use \code{modelFormula_random = 
"(1|plate_ID) + (1|participant_ID:plate_ID)"}.
See \code{?lmer()}.}

\item{reduced_modelFormula_fixed}{Optional reduced model formula 
for fixed effects that contains only a subset of the terms in modelFormula. 
This could be an empty string if the full model contains only one term.
The reduced model serves as null model for a likelihood ratio test 
(LRT, which is a Chi-square test) using \code{anova()}. 
This could be useful for testing the overall significance of factor 
variables with more than 2 levels, for example, testing the overall significance 
of a categorical time effect. The reduced model uses the same random effects 
as specified in \code{modelFormula_random}.}

\item{exclude_targets}{A vector of target names for targets that will be 
excluded from the differential expression analysis. Internal control targets, 
for example, should probably always be excluded.}

\item{exclude_samples}{A vector of sample names for samples that will be excluded
from differential expression analysis. External control wells (IPCs, NCs, SC,)
should usually be excluded.}

\item{target_subset}{Overrides exclude_targets. A vector of target names 
for targets that will be included in differential expression analysis.}

\item{sample_subset}{Overrides exclude_samples. A vector of sample names 
for samples that will be included in differential expression analysis.}

\item{return_model_fits}{Logical \code{TRUE} or \code{FALSE} (default).
Should a list of the model fits be returned? Might be useful for more 
detailed analyses and plotting. However, also requires using more memory.}
}
\value{
A list including the following:
\item{modelStats}{A data frame with rows corresponding to targets and columns 
corresponding to estimated model coefficients, unadjusted p-values, 
Bonferroni adjusted p-values, and Benjamini-Hochberg false discovery rate
adjusted p-values (see \code{?p.adjust()})}
\item{modelFits}{A list of length equal to number of targets containing
the model fit output from \code{lm()}. Only returned when 
\code{return_model_fits=TRUE}.}
\item{LRTstats}{A data frame with rows corresponding to targets and columns }
}
\description{
Fits linear mixed effects model to each target in the NULISAseq data set. 
Outputs coefficients, t-statistics, unadjusted and adjusted p-values.
Uses \code{lme4} and \code{lmerTest} packages.
}
