% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNULISAseq.R
\name{loadNULISAseq}
\alias{loadNULISAseq}
\title{Convenience function for use by SAM
Read NULISAseq XML, perform normalization, and QC}
\usage{
loadNULISAseq(file, IPC, IC, ...)
}
\arguments{
\item{file}{Character string. Path and name of the file.}

\item{IPC}{String(s) that represent the inter-plate control wells. 
For example, \code{'IPC'} (default). Set to \code{NULL} if there are no IPCs.
Only used for xml file formats.}

\item{IC}{string(s) that represents the names of internal control targets. 
Default is \code{'mCherry'}.Only used for xml file formats.}
}
\value{
List of lists, data frames, and matrices.
Output will differ slightly depending on the input file type.
}
\description{
Reads NULISAseq XML file, where XML file is output from the Alamar
Biosciences Galaxy NULISAseq (Beta) tool or the NULISAseq Normalization (Alpha) tool.
}
