% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNULISAseq.R
\name{loadNULISAseq}
\alias{loadNULISAseq}
\title{Read NULISAseq XML, perform normalization, and QC}
\usage{
loadNULISAseq(
  file,
  IC = NULL,
  IPC = NULL,
  SC = NULL,
  NC = NULL,
  TAP = TRUE,
  Bridge = NULL,
  sample_group_covar = "SAMPLE_MATRIX",
  plateID = NULL,
  scaleFactor = 10^4,
  transformReverse_scaleFactor = 10^4,
  replace_cal_blank_zeros = FALSE,
  replace_zeros_with_NA = TRUE,
  security = TRUE,
  excludeSamples = NULL,
  excludeTargets = NULL,
  advancedQC = FALSE,
  ...
)
}
\arguments{
\item{file}{Character string. Path and name of the file.}

\item{IC}{string(s) that represents the names of internal control targets. 
Default is \code{'mCherry'}.Only used for xml file formats.}

\item{IPC}{string(s) present in the sample names
that represent the inter-plate control wells. 
For example, \code{'IPC'}. Set to \code{NULL} (default) to use the 
type variable from the Barcode B file or if there are no IPCs.
Only used for xml file formats.}

\item{SC}{string(s) present in the sample names
that represent the sample control wells. 
For example, \code{'SC'}. Set to \code{NULL} (default) to use the 
type variable from the Barcode B file or if there are no IPCs.
Only used for xml file formats.}

\item{NC}{string(s) present in the sample names
that represent the negative control wells. 
For example, \code{'NC'}. Set to \code{NULL} (default) to use the 
type variable from the Barcode B file or if there are no NCs.
Only used for xml file formats.}

\item{TAP}{If TRUE (default), uses TAP detectability criteria in sample QC 
which includes more matrix types than non-TAP criteria.}

\item{Bridge}{string(s) present in the sample names
that represent the bridge sample wells. 
Set to \code{NULL} (default) to use the 
type variable from the Barcode B file or 
if there are no bridge samples (default).
Only used for xml file formats.}

\item{sample_group_covar}{Optional column name in the Barcode B file 
and samples data matrix output by readNULISAseq that represents subgroups
for which detectability will be calculated separately, in addition to 
overall detectability. Default is 'SAMPLE_MATRIX', Function will check first to
be sure that the variable is present in the column names of the samples matrix.
Can be set to NULL to not use this feature.}

\item{replace_cal_blank_zeros}{Logical TRUE / FALSE. Default is FALSE.
This parameter is passed to the applyAQ function.
If FALSE, the "a" parameter for targets with a zero blank calibrator (NC) mean 
is set using the blank calibrator mean value (which would equal zero in these cases).
IF TRUE, the "a" parameter for targets with a zero blank calibrator mean 
is set using the nonzero master curve "a" parameter 
estimate instead.}

\item{replace_zeros_with_NA}{Logical TRUE / FALSE. 
This parameter is passed to the applyAQ function.
When TRUE (default), any zero values 
in the AQ data output will be replaced with NA. When FALSE, these values remain as zero.}

\item{security}{Logical. Default is TRUE. Should security checks be performed 
before generating AQ data.}

\item{excludeSamples}{A vector of sample names that will be excluded from all outputs.}

\item{excludeTargets}{A vector of target names that will be excluded from all outputs.}

\item{advancedQC}{Whether to use advancedQC metrics}
}
\value{
List of lists, data frames, and matrices.
Output will differ slightly depending on the input file type.
}
\description{
Reads NULISAseq XML file
}
