% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod.R
\name{lod}
\alias{lod}
\title{Calculate Limits of Detection}
\usage{
lod(data_matrix, blanks, min_count = 0)
}
\arguments{
\item{data_matrix}{The Data matrix output from readNULISAseq.R
or normalized data from normalization functions.}

\item{blanks}{Column indices or column names of the blanks in the
data_matrix.}

\item{min_count}{Optional count threshold to apply in addition
to the LoD. Default is 0.}

\item{LOD}{Vector of limits of detection.}

\item{aboveLOD}{Logical matrix indicating whether counts are 
above or below LoD for that target.}
}
\value{
A list.
}
\description{
Calculates limit of detection (LoD) for each target based on the
negative controls (blanks). LoD = mean(blanks) + 3*SD(blanks).
Designates data as either above or below LoD. 
Option to specify minimum count threshold for detectability.
}
