% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod.R
\name{lod}
\alias{lod}
\title{Calculate Limits of Detection}
\usage{
lod(
  data_matrix,
  blanks,
  min_count = 0,
  min_blank_no = 4,
  mad_threshold = 2.5,
  ignore_target_blank = NULL,
  targetNoOutlierDetection = NULL,
  match_matrix = NULL
)
}
\arguments{
\item{data_matrix}{The Data matrix output from readNULISAseq.R
or normalized data from normalization functions.}

\item{blanks}{Column indices or column names of the blanks in the
data_matrix.}

\item{min_count}{Optional count threshold to apply in addition
to the LoD. Default is 0.}

\item{min_blank_no}{Optional numeric parameter defining the minimum number of
blanks required to enable MAD outlier detection and removal. Default is 4.}

\item{mad_threshold}{Optional numeric parameter defining the threshold used for
MAD outlier identification. Default is 2.5.}

\item{ignore_target_blank}{List of targets and corresponding blanks/NCs to exclude.
Names of the list represent targets and
values which are arrays represent blank/NC names to ignore during LoD calculation.}

\item{targetNoOutlierDetection}{Option to provide targets which should NOT have 
outlier detection applied
Defaults to NULL.}

\item{match_matrix}{Matrix of indices provided by calcSampleTargetNAs. 
Lists samples/targets that should not be reported}
}
\value{
A list containing:
\item{LOD}{Vector of limits of detection.}
\item{aboveLOD}{Logical matrix indicating whether counts are 
above or below LoD for that target.}
}
\description{
Calculates limit of detection (LoD) for each target based on the
negative controls (blanks). LoD = mean(blanks) + 3*SD(blanks).
LoD is typically calculated on IC-IPC normalized reads (unlogged).
Designates data as either above or below LoD. 
Option to specify minimum count threshold for detectability.
}
