% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULISAseqR.R
\docType{_PACKAGE}
\name{NULISAseqR}
\alias{NULISAseqR}
\title{NULISAseqR Package}
\description{
NULISAseqR package can load data from XML data files.
Includes functions to normalize, calculate CV, and visualize data, among others.
}
\author{
Alamar Biosciences \email{jbeer@alamarbio.com}
}
