% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateHeatmap.R
\name{plateHeatmap}
\alias{plateHeatmap}
\title{Draw Plate Heatmap}
\usage{
plateHeatmap(
  target_data,
  well_order = NULL,
  print_counts = TRUE,
  cex = 0.6,
  cex.axis = 0.8,
  digits = 0,
  title = NULL,
  relative = F
)
}
\arguments{
\item{target_data}{A numeric vector of length 96 that represents the 
counts that will be shown on the heatmap.}

\item{well_order}{Optional. A numeric vector used to order the wells 
from A01-A12, B01-B12, ..., H01-H12. If not provided, then 
it is assumed that the target_data is already in this order.}

\item{print_counts}{A TRUE / FALSE that specifies whether or not 
counts will be printed in each cell of the heatmap. Default is TRUE.}

\item{cex}{Character expansion factor that controls size of the 
printed counts. Default is 0.7.}

\item{cex.axis}{Character expansion factor that controls size of the 
axis labels. Default is 0.8.}

\item{digits}{number of digits after the decimal point to print. 
Default is 0.}

\item{title}{Optional. Plot title.}
}
\value{
Draws a heatmap.
}
\description{
Makes heatmap for a specified target or total counts.
}
