% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC2XML}
\alias{QC2XML}
\title{Write Processed XML from QC table}
\usage{
QC2XML(input, QCNode, type = "plate", combined = F)
}
\arguments{
\item{input}{QC table}

\item{type}{whether the table is sample, target or plate}

\item{combined}{}

\item{QC}{QC XML node to add to}
}
\value{
QC XML
}
\description{
Writes NULISAseq QC flags to XML
}
\examples{
# QC2XML(inputtable)

}
