% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCFlagPlate}
\alias{QCFlagPlate}
\title{Write Plate QC table}
\usage{
QCFlagPlate(
  raw,
  normed,
  targets,
  samples,
  ICs = NULL,
  IPCs = NULL,
  NCs = NULL,
  SCs = NULL
)
}
\arguments{
\item{raw}{unnormalized data}

\item{ICs}{indices of IC samples}

\item{IPCs}{indices of IPC samples}

\item{NCs}{indices of NC samples}

\item{SCs}{indices of SC samples}

\item{normednormalized}{data}
}
\value{
QC table
}
\description{
Writes NULISAseq QC plate flags
}
\examples{
# QCFlagPlate(inputtable)

}
