% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCFlagPlate}
\alias{QCFlagPlate}
\title{Write Plate QC table}
\usage{
QCFlagPlate(
  raw,
  normed,
  aboveLOD,
  targets,
  samples,
  ICs = NULL,
  IPCs = NULL,
  NCs = NULL,
  SCs = NULL,
  AQ = TRUE,
  AQ_QC = NULL,
  Sample_QC = NULL
)
}
\arguments{
\item{raw}{unnormalized data}

\item{normed}{normalized data}

\item{aboveLOD}{aboveLOD from lod function}

\item{ICs}{names of IC samples}

\item{IPCs}{names of IPC samples}

\item{NCs}{names of NC samples}

\item{SCs}{names of SC samples}

\item{AQ}{Whether this assay is AQ}

\item{AQ_QC}{AQ QC metrics, these are necessary to calculate a plate QC metric for AQ}

\item{Sample_QC}{Sample QC metrics, one of these is necessary to calculate a plate QC metric for AQ}
}
\value{
QC table
}
\description{
Writes NULISAseq QC plate flags
}
\examples{
# QCFlagPlate(inputtable)

}
