% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCFlagPlate}
\alias{QCFlagPlate}
\title{Write Plate QC table}
\usage{
QCFlagPlate(raw, normed, ICs, NCs, IPCs)
}
\arguments{
\item{raw}{unnormalized data}

\item{ICs}{indices of IC samples}

\item{NCs}{indices of NC samples}

\item{IPCs}{indices of IPC samples}

\item{normednormalized}{data}
}
\value{
QC table
}
\description{
Writes NULISAseq QC plate flags
}
\examples{
# QCFlagPlate(inputtable)

}
