% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCFlagSample}
\alias{QCFlagSample}
\title{Write Sample QC table}
\usage{
QCFlagSample(raw, normed, ICs, NCs, IPCs, samples, well_order = NULL)
}
\arguments{
\item{raw}{unnormalized data}

\item{ICs}{indices of IC samples}

\item{NCs}{indices of NC samples}

\item{IPCs}{indices of IPC samples}

\item{normednormalized}{data}
}
\value{
QC table
}
\description{
Writes NULISAseq QC sample flags
}
\examples{
# QCFlagSample(inputtable)

}
