% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCFlagTarget}
\alias{QCFlagTarget}
\title{Write Target QC table}
\usage{
QCFlagTarget(
  AQdata,
  raw,
  IPCnormed,
  detectability,
  aboveLOD,
  withinDR,
  absRun,
  targets,
  samples,
  SCparams,
  ICs = NULL,
  IPCs = NULL,
  NCs = NULL,
  SCs = NULL,
  advancedQC = FALSE
)
}
\arguments{
\item{AQdata}{Absolute quantification data}

\item{raw}{Raw data}

\item{IPCnormed}{IPC-normalized data}

\item{detectability}{detectability}

\item{aboveLOD}{aboveLOD}

\item{withinDR}{withinDR}

\item{absRun}{whether the assay is absolute quantification}

\item{targets}{targets data}

\item{samples}{samples data}

\item{SCparams}{SC concentrations}

\item{ICs}{names of IC samples}

\item{IPCs}{names of IPC samples}

\item{NCs}{names of NC samples}

\item{SCs}{names of SC samples}
}
\value{
QC table
}
\description{
Writes NULISAseq QC target flags
}
\examples{
# QCFlagTarget(inputtable)

}
