% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNULISAseq.R
\name{readNULISAseq}
\alias{readNULISAseq}
\title{Read NULISAseq XML}
\usage{
readNULISAseq(
  xml_file,
  plateID = NULL,
  file_type = "xml_no_mismatches",
  replaceNA = TRUE
)
}
\arguments{
\item{xml_file}{Character string. Path and name of the file.}

\item{plateID}{Character string that denotes plate ID.}

\item{file_type}{Character string. Type of input file, as output from Galaxy. Options include
xml_full_output, xml_no_mismatches (default) (both from NULISAseq tool),
or xml_normalization (from NULISAseq Normalization tool).}

\item{replaceNA}{Logical. If TRUE (default), will replace missing counts with 
zeros.}
}
\value{
List of lists, data frames, and matrices.
Output will differ slightly depending on the input file type.
}
\description{
Reads NULISAseq XML file, where XML file is output from the Alamar
Biosciences Galaxy NULISAseq (Beta) tool or the NULISAseq Normalization (Alpha) tool.
}
