% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNULISAseq.R
\name{readNULISAseq}
\alias{readNULISAseq}
\title{Read NULISAseq XML}
\usage{
readNULISAseq(
  file,
  plateID = NULL,
  file_type = "xml_no_mismatches",
  target_column_names = NULL,
  sample_column_names = NULL,
  sample_group_covar = NULL,
  IC = NULL,
  IPC = NULL,
  SC = NULL,
  NC = NULL,
  Bridge = NULL,
  Calibrator = NULL,
  replaceNA = TRUE
)
}
\arguments{
\item{file}{Character string. Path and name of the file.}

\item{plateID}{Character string that denotes plate ID.}

\item{file_type}{Character string. Type of input file.
Options include \code{xml_no_mismatches}, \code{xlsx} (output from ACC)
or \code{csv_long} (output from \code{writeNULISAseq()} function).}

\item{target_column_names}{A vector of column names for target-specific data
that will appear in the target data frame output. 
Only used for \code{csv_long} format, and only needed when target column names
are different from the default, which are \code{c('Target', 'AlamarTargetID', 
'UniProtID', 'ProteinName')}.}

\item{sample_column_names}{A vector of column names for sample-specific data
that will appear in the sample data frame output. 
Only used for \code{csv_long} format, and only needed when sample column names
are different from the default. Default includes all column names in the csv except 
for the target column names, given above (either the default or whatever is specified), 
and the following: \code{c('Panel', 'PanelLotNumber', 'LOD', 'NPQ')}. 
The sample-specific data
will include by default \code{c('PlateID', 'SampleName', 'SampleType', 'SampleQC')}.
In addition, by default, it will include any other variables in the dataset.
If \code{sample_column_names} is specified, however, it will only include
those variables in the sample data frame output.}

\item{sample_group_covar}{string that represents the name of the covariate
that defines sample group information. Defaults to NULL. Typically, this covariate
defines whether each sample is plasma, serum, csf, etc.
This information can be used when reporting detectability. For instance report sample
group specific detectability values, etc.).}

\item{IC}{string(s) that represents the names of internal control targets. 
Default is \code{'mCherry'}.Only used for xml file formats.}

\item{IPC}{string(s) present in the sample names
that represent the inter-plate control wells. 
For example, \code{'IPC'}. Set to \code{NULL} (default) to use the 
type variable from the Barcode B file or if there are no IPCs.
Only used for xml file formats.}

\item{SC}{string(s) present in the sample names
that represent the sample control wells. 
For example, \code{'SC'}. Set to \code{NULL} (default) to use the 
type variable from the Barcode B file or if there are no IPCs.
Only used for xml file formats.}

\item{NC}{string(s) present in the sample names
that represent the negative control wells. 
For example, \code{'NC'}. Set to \code{NULL} (default) to use the 
type variable from the Barcode B file or if there are no NCs.
Only used for xml file formats.}

\item{Bridge}{string(s) present in the sample names
that represent the bridge sample wells. 
Set to \code{NULL} (default) to use the 
type variable from the Barcode B file or 
if there are no bridge samples (default).
Only used for xml file formats.}

\item{Calibrator}{string(s) present in the sample names
that represent the calibrator wells. 
Set to \code{NULL} (default) to use the 
type variable from the Barcode B file or 
if there are no calibrator samples (default).
Only used for xml file formats.}

\item{replaceNA}{Logical. If TRUE (default), will replace missing counts with 
zeros.}
}
\value{
List of lists, data frames, and matrices.
Output will differ slightly depending on the input file type.
}
\description{
Reads NULISAseq XML file
}
