% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetQCplot.R
\name{targetQCplot}
\alias{targetQCplot}
\title{Generate a Target QC Plot}
\usage{
targetQCplot(runs)
}
\arguments{
\item{runs}{A named list of run data output from \code{laodNULISAseq()} function 
or a list of these outputs for multiple runs.}
}
\value{
Generates a target QC plot.
}
\description{
Generates a quantifiability histogram using output of 
\code{quantifiability()} function.
}
