% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano Plot}
\usage{
volcanoPlot(estimate, pval, title, xlimits = NULL, ylimits = NULL)
}
\value{
Draws a plot.
}
\description{
Makes volcano plot using output of deTest function.
}
