% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot for NULISAseq differential expression test}
\usage{
volcanoPlot(
  coefs,
  p_vals,
  target_labels,
  title = NULL,
  xlabel = expression("log"[2] * "(fold change)"),
  ylabel = expression("-log"[10] * "(FDR-adjusted p-value)"),
  xlimits = NULL,
  ylimits = NULL,
  sig_threshold = 0.05,
  sig_label = "FDR = 5\%",
  label_all_targets = FALSE,
  target_labels_off = FALSE,
  target_label_colors = NULL,
  target_point_colors = NULL,
  target_label_size = 2,
  target_label_segment_color = "grey",
  max.overlaps = Inf,
  plot_title_font_size = 14,
  axis_label_font_size = 12,
  tick_label_font_size = 12,
  plot_aspect_ratio = 1,
  log_y = TRUE
)
}
\arguments{
\item{coefs}{Vector of coefficients.}

\item{p_vals}{Vector of p-values.}

\item{target_labels}{Vector of target names.}

\item{title}{Plot title.}

\item{xlabel}{X axis label.}

\item{ylabel}{Y axis label.}

\item{xlimits}{X axis limits}

\item{ylimits}{Y axis limits}

\item{sig_threshold}{P-value significance threshold.}

\item{sig_label}{Label to mark significance threshold on plot.}

\item{label_all_targets}{Logical TRUE or FALSE. 
Should all targets be labelled? Default is FALSE 
and will only label significant targets.}

\item{target_labels_off}{Logical TRUE or FALSE. Should no targets be labelled?}

\item{target_label_colors}{A vector of the same length as number of targets
and same order as targets with colors for the target labels. Default is to 
color upregulated target labels red and downregulated target labels blue.}

\item{target_point_colors}{A vector of the same length as number of targets
and same order as targets with colors for the points. Default is to 
make significant upregulated target points red, significant downregulated 
target points blue, and nonsignificant target points grey.}

\item{target_label_size}{Font size of target labels.}

\item{target_label_segment_color}{Color of line segments for target labels.
Can be a single value or a vector same length as number of targets.}

\item{max.overlaps}{Passed to ggrepel. Integer which determines how many targets 
will have labels. "Inf" (default) labels all targets.}

\item{plot_title_font_size}{Font size of plot title.}

\item{axis_label_font_size}{Font size of axis title.}

\item{tick_label_font_size}{Font size of axis tick marks.}

\item{plot_aspect_ratio}{Aspect ratio for plot. Default is 1 (square plot).}

\item{log_y}{Logical \code{TRUE} (default) or \code{FALSE}. Should y-axis be 
-log10 transformed? (\code{TRUE} recommended for plotting p-values.)}
}
\value{
Outputs a ggplot object.
}
\description{
Draws a volcano plot for a set of coefficients and p-values. Uses ggplot
for plotting and ggrepel package for labels. By default, plot labeling
assumes FDR-corrected p-values.
}
